finally working.
but it doesnt smartly choose the best move, due to the fact that there is no difference between all +1, so it tends to choose the first available move, not the one that give shortest winning.

possible solution is to modify the utility function / implement the evaluation function.

also, only works for player 1 only, not as the second movement